refactor(materializer): extract interaction-index-patch resolvers (file #3)
Lift the incremental interaction-index / transcript-entry-index patch resolvers
into interaction-index-patch.js. Net 74/74, check 0.
